+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 4.24.3.0
+
+- Support `table` `items` data-binding of table row cell `background`, `foreground`, `font`, `image` to model attributes by convention based on specified `column_properties` (e.g. for `column_properties` model `name` attribute, automatically assume model `name_background` provides `[r, g, b]` array of color)
+- Update Hello, Table! sample to demonstrate data-binding `table` `items` `background`, `foreground`, `font`, `image` to model attributes by convention
+
 ## 4.24.2.3
 
 - Default `code_text` font name in Mac is changed back to `'Courier'`